Remove HyperDhtInner & SO MUCH CODE
This made SO MUCH code obsolete
