+++ hdht/src/lib.rs
@@ -3,18 +3,10 @@
 #![deny(clippy::enum_glob_use)]
 
 use std::{
-    any::Any,
     array::TryFromSliceError,
-    cmp,
-    collections::BTreeMap,
-    convert::{TryFrom, TryInto},
     fmt,
-    future::Future,
-    net::{AddrParseError, IpAddr, SocketAddr, SocketAddrV4, ToSocketAddrs},
-    ops::DerefMut,
-    pin::Pin,
-    sync::{Arc, Mutex, RwLock},
-    time::Duration,
+    net::{AddrParseError, SocketAddr, SocketAddrV4},
+    sync::Arc,
 };
 
 use cenc::{
@@ -22,47 +14,18 @@ use cenc::{
     RelayThroughInfoBuilderError, UdxInfoBuilderError,
 };
 use compact_encoding::{CompactEncoding, EncodingError};
-use crypto::PublicKey;
-use dht_rpc::{
-    cenc::generic_hash,
-    commit::{Commit, CommitMessage, CommitRequestParams, Progress},
-    io::{InResponse, MessageSender, OutRequestBuilder},
-    query::{CommandQuery, Query, QueryId, QueryResult as RpcQueryResult},
-    Bootstrapped, Command, DhtConfig, IdBytes, Peer, PeerId, RequestFutureError, RequestMsgData,
-    RequestOk, ResponseOk, RpcDht, RpcDhtBuilderError, RpcDhtEvent, Tid,
-};
-use futures::{
-    channel::mpsc::{self},
-    task::{Context, Poll},
-    Stream, StreamExt,
-};
-use futuresmap::FuturesMap;
-use prost::Message as ProstMessage;
-use queries::{
-    AnnounceClearResult, AnnounceInner, AunnounceClearInner, FindPeerInner, FindPeerResponse,
-    LookupInner, LookupResponse, QueryResult, UnannounceInner, UnannounceResult,
-};
-use smallvec::alloc::collections::VecDeque;
-use tokio::sync::oneshot::{self, error::RecvError};
-use tracing::{debug, error, instrument, trace, warn};
+use dht_rpc::{io::InResponse, IdBytes, RequestFutureError, RpcDhtBuilderError};
+use tokio::sync::oneshot::error::RecvError;
+use tracing::{error, warn};
 
-use crate::{
-    cenc::HandshakeSteps,
-    dht_proto::{PeersInput, PeersOutput},
-    lru::{CacheKey, PeerCache},
-    next_router::connection::Connection,
-    store::Store,
-};
+use crate::cenc::HandshakeSteps;
 
 mod dht_proto {
     include!(concat!(env!("OUT_DIR"), "/dht_pb.rs"));
 }
-pub mod cenc;
+mod cenc;
 mod crypto;
-mod futuresmap;
-pub mod lru;
 mod next_router;
-mod queries;
 mod store;
 
 pub mod adht;
@@ -160,609 +123,6 @@ impl From<EncodingError> for Error {
     }
 }
 
-/// The implementation of the hyperswarm DHT
-#[derive(Debug)]
-pub struct HyperDhtInner {
-    /// The underlying Rpc DHT including IO
-    rpc: RpcDht,
-    /// Map to track the queries currently in progress
-    queries: FuturesMap<QueryId, QueryStreamType>,
-    /// If `true`, the node will become non-ephemeral after the node has shown, to be long-lived
-    #[expect(unused)] // FIXME why aint this used
-    adaptive: bool,
-    /// Cache for known peers
-    peers: PeerCache,
-    /// Storage for the mutable/immutable values
-    store: Store,
-    /// Queued events to return when being polled.
-    queued_events: VecDeque<HyperDhtEvent>,
-    /// Default keypair
-    #[expect(unused)] // TODO
-    default_keypair: Keypair,
-    /// Router for peer connections
-    router: next_router::Router,
-    /// Channels to send FindPeerResponse through for pending find_peer queries
-    pending_find_peers: BTreeMap<QueryId, oneshot::Sender<Arc<QueryResult>>>,
-    /// inflight
-    pending_msg: BTreeMap<Tid, Box<dyn Any + Send>>,
-}
-
-impl HyperDhtInner {
-    /// Create a new DHT based on the configuration
-    pub async fn with_config(mut config: DhtConfig) -> Result<Self> {
-        if config.bootstrap_nodes.is_empty() {
-            for addr_str in DEFAULT_BOOTSTRAP.iter() {
-                if let Some(addr) = addr_str.to_socket_addrs()?.last() {
-                    config.bootstrap_nodes.push(addr)
-                }
-            }
-        }
-
-        Ok(Self {
-            adaptive: config.adaptive,
-            queries: Default::default(),
-            rpc: RpcDht::with_config(config).await?,
-            // peer cache with 25 min timeout
-            peers: PeerCache::new(65536, Duration::from_secs(60 * 25)),
-            store: Store::new(5000),
-            queued_events: Default::default(),
-            default_keypair: Default::default(),
-            router: Default::default(),
-            pending_find_peers: Default::default(),
-            pending_msg: Default::default(),
-        })
-    }
-
-    /// The local address of the underlying `UdpSocket`
-    #[inline]
-    pub fn local_addr(&self) -> Result<SocketAddr> {
-        Ok(self.rpc.local_addr()?)
-    }
-
-    /// Handle an incoming requests for the registered commands and reply.
-    fn on_command(&mut self, q: CommandQuery, request: RequestMsgData, peer: Peer) {
-        match q.command {
-            MUTABLE_STORE_CMD => {
-                let resp = self.store.on_command_mut(q);
-                self.rpc.reply_command(resp)
-            }
-            IMMUTABLE_STORE_CMD => {
-                let resp = self.store.on_command(q);
-                self.rpc.reply_command(resp)
-            }
-            PEERS_CMD => self.on_peers(q),
-            command => self
-                .queued_events
-                .push_back(HyperDhtEvent::CustomCommandQuery {
-                    command,
-                    msg: Box::new(request),
-                    peer,
-                }),
-        }
-    }
-    /// Callback for an incoming `peers` command query
-    fn on_peers(&mut self, mut query: CommandQuery) {
-        // decode the received value
-        if let Some(ref val) = query.value {
-            if let Ok(peer) = PeersInput::decode(&**val) {
-                // callback
-                let port = peer
-                    .port
-                    .and_then(|port| u16::try_from(port).ok())
-                    .unwrap_or_else(|| query.peer.addr.port());
-
-                if let IpAddr::V4(host) = query.peer.addr.ip() {
-                    let from = SocketAddr::V4(SocketAddrV4::new(host, port));
-
-                    let remote_cache = CacheKey::Remote(query.target);
-
-                    let local_cache = peer.local_address.as_ref().and_then(|l| {
-                        if l.len() == 6 {
-                            let prefix: [u8; 2] = l[0..2].try_into().unwrap();
-                            let suffix: [u8; 4] = l[2..].try_into().unwrap();
-                            Some((
-                                CacheKey::Local {
-                                    id: query.target,
-                                    prefix,
-                                },
-                                suffix,
-                            ))
-                        } else {
-                            None
-                        }
-                    });
-
-                    let local_peers = if let Some((local_cache, suffix)) = local_cache {
-                        self.peers.get(&local_cache).and_then(|addrs| {
-                            addrs.iter_locals().map(|locals| {
-                                locals
-                                    .filter(|s| **s != suffix)
-                                    .flat_map(|s| s.iter())
-                                    .cloned()
-                                    .take(32)
-                                    .collect::<Vec<_>>()
-                            })
-                        })
-                    } else {
-                        None
-                    };
-
-                    let peers = if let Some(remotes) = self
-                        .peers
-                        .get(&remote_cache)
-                        .and_then(|addrs| addrs.remotes())
-                    {
-                        let num = cmp::min(
-                            remotes.len(),
-                            128 - local_peers.as_ref().map(|l| l.len()).unwrap_or_default(),
-                        );
-                        let mut buf = Vec::with_capacity(num * 6);
-
-                        for addr in remotes.iter().filter(|addr| **addr != from).take(num) {
-                            if let IpAddr::V4(ip) = addr.ip() {
-                                buf.extend_from_slice(&ip.octets()[..]);
-                                buf.extend_from_slice(&addr.port().to_be_bytes()[..]);
-                            }
-                        }
-                        Some(buf)
-                    } else {
-                        None
-                    };
-
-                    let output = PeersOutput { peers, local_peers };
-                    let mut buf = Vec::with_capacity(output.encoded_len());
-
-                    // fits safe in vec
-                    output.encode(&mut buf).unwrap();
-                    query.value = Some(buf);
-                    self.rpc.reply_command(query.into());
-                    return;
-                }
-                let _ = query.value.take();
-                self.rpc.reply_command(query.into());
-            }
-        }
-    }
-
-    pub fn find_peer(&mut self, pub_key: PublicKey) -> QueryId {
-        let target = IdBytes(generic_hash(&*pub_key));
-
-        let query_id = self
-            .rpc
-            .query(commands::FIND_PEER, target, None, Commit::No);
-        self.queries.insert(
-            query_id,
-            QueryStreamType::FindPeer(FindPeerInner::new(query_id, target)),
-        );
-        query_id
-    }
-
-    /// Initiates an iterative query to the closest peers to lookup the topic.
-    ///
-    /// The result of the query is delivered in a
-    /// [`HyperDhtEvent::LookupResult`].
-    pub fn lookup(&mut self, target: IdBytes, commit: Commit) -> QueryId {
-        let query_id = self.rpc.query(commands::LOOKUP, target, None, commit);
-        self.queries.insert(
-            query_id,
-            QueryStreamType::Lookup(LookupInner::new(query_id, target)),
-        );
-        query_id
-    }
-
-    /// Announce the topic to the closest peers
-    ///
-    /// Query result is a [`HyperDhtEvent::AnnounceResult`].
-    pub fn announce(
-        &mut self,
-        target: IdBytes,
-        key_pair: &Keypair,
-        _relay_addresses: &[SocketAddr],
-    ) -> QueryId {
-        let qid = self.rpc.query(
-            commands::LOOKUP,
-            target,
-            None,
-            Commit::Custom(Progress::default()),
-        );
-        self.queries.insert(
-            qid,
-            QueryStreamType::Announce(AnnounceInner::new(qid, target, key_pair.clone())),
-        );
-        qid
-    }
-
-    /// Announce the topic to the closest peers, send unnanounce while doing so
-    ///
-    /// Query result is a [`HyperDhtEvent::AnnounceResult`].
-    pub fn announce_clear(
-        &mut self,
-        target: IdBytes,
-        key_pair: &Keypair,
-        _relay_addresses: &[SocketAddr],
-    ) -> QueryId {
-        let qid = self.rpc.query(
-            commands::LOOKUP,
-            target,
-            None,
-            Commit::Custom(Progress::default()),
-        );
-        self.queries.insert(
-            qid,
-            QueryStreamType::AnnounceClear(AunnounceClearInner::new(target, key_pair.clone())),
-        );
-        qid
-    }
-
-    /// Initiates an iterative query to unannounce the topic to the closest
-    /// peers.
-    ///
-    /// The result of the query is delivered in a
-    /// [`HyperDhtEvent::UnAnnounceResult`].
-    pub fn unannounce(&mut self, target: IdBytes, key_pair: &Keypair) -> QueryId {
-        let qid = self.rpc.query(commands::LOOKUP, target, None, Commit::No);
-        self.queries.insert(
-            qid,
-            QueryStreamType::UnAnnounce(UnannounceInner::new(target, key_pair.clone())),
-        );
-        qid
-    }
-
-    #[instrument(skip_all)]
-    fn inject_query_response(
-        &mut self,
-        resp: Arc<InResponse>,
-        _cx: &mut Context<'_>,
-    ) -> Result<()> {
-        // Holding `&mut query` here prevents us from calling self.request_unannounce
-        // maybe instead we return something from the block to do the msg that we want?
-        // however I want to pass in an id from the request.
-        // I really just want a mut ref to the query, but
-        //
-        let Some((query, qid)) = resp
-            .query_id
-            .and_then(|qid| self.queries.get_mut(&qid).map(|q| (q, qid)))
-        else {
-            error!("query response without matching query!");
-            return Ok(());
-        };
-        let event = {
-            match query.deref_mut() {
-                QueryStreamType::Announce(inner) => {
-                    inner.inject_response(resp);
-                    None
-                }
-                QueryStreamType::UnAnnounce(inner) => {
-                    inner.inject_response(&mut self.rpc.io, resp, qid);
-                    None
-                }
-                QueryStreamType::Lookup(inner) => inner.inject_response(resp.clone()),
-                QueryStreamType::AnnounceClear(inner) => {
-                    inner.inject_response(&mut self.rpc.io, resp, qid);
-                    None
-                }
-                QueryStreamType::FindPeer(inner) => inner
-                    .inject_response(resp.clone())
-                    .map(HyperDhtEvent::FindPeerResponse),
-            }
-        };
-        if let Some(e) = event {
-            self.queued_events.push_back(e);
-        }
-        Ok(())
-    }
-
-    #[instrument(skip_all)]
-    fn inject_response(&mut self, resp: Arc<InResponse>, _cx: &mut Context<'_>) -> Result<()> {
-        trace!(
-            cmd = display(resp.cmd()),
-            "Handle Response for custom command"
-        );
-        // handle respones to non-queries
-        match resp.request.command {
-            Command::Internal(_) => {
-                // TODO
-            }
-            commands::PEER_HANDSHAKE => {
-                let phs_resp = decode_peer_handshake_response(&resp)?;
-                // self.queued_events
-                //     .push_back(HyperDhtEvent::PeerHandshakeResponse(phs_resp.clone()));
-
-                let tx_opt = self
-                    .pending_msg
-                    .remove(&resp.request.tid)
-                    .and_then(|boxed| boxed.downcast::<oneshot::Sender<Result<Connection>>>().ok());
-                let conn = self
-                    .router
-                    .inject_response(&resp, phs_resp, self.rpc.socket())
-                    .inspect_err(|e| error!("Error handling peer handshake response: [{e}]"))?;
-                tx_opt.map(|tx| tx.send(Ok(conn.clone())));
-
-                self.queued_events
-                    .push_back(HyperDhtEvent::Connected((resp.request.tid, conn)));
-            }
-            Command::External(_) => {
-                // TODO
-            }
-        }
-        Ok(())
-    }
-
-    // A query was completed
-    fn query_target_search_done(&mut self, query_result: Arc<RpcQueryResult>) {
-        if let Some(query) = self.queries.get_mut(&query_result.query_id) {
-            query.target_search_done(self.rpc.io.create_sender(), query_result);
-        } else {
-            warn!(
-                id = ?query_result.query_id,
-                "Query with unknown id finished"
-            );
-        }
-    }
-
-    fn commit(&mut self, query: Arc<RwLock<Query>>, channel: mpsc::Sender<CommitMessage>) {
-        let id = query.read().unwrap().id();
-        let Some(qst) = self.queries.get_mut(&id) else {
-            error!("Tried to commit with an unknown query id: [{id:?}]");
-            panic!("Tried to commit with an unknown query id: [{id:?}]");
-        };
-        qst.commit(query, channel);
-    }
-
-    pub fn request_peer_handshake(
-        &mut self,
-        remote_public_key: PublicKey,
-        destination: SocketAddrV4,
-    ) -> Result<Tid> {
-        let tid = self.rpc.new_tid();
-
-        let SocketAddr::V4(addr) = self.local_addr()? else {
-            todo!()
-        };
-
-        let value = self
-            .router
-            .first_step(tid, *remote_public_key, Some(vec![addr]), self.rpc.socket())
-            .unwrap();
-        let req = OutRequestBuilder::new(Peer::new(destination.into()), commands::PEER_HANDSHAKE)
-            .value(value)
-            .target(generic_hash(&*remote_public_key).into())
-            .tid(tid);
-        self.rpc.request(req);
-        Ok(tid)
-    }
-
-    #[expect(clippy::too_many_arguments)]
-    fn request_announce_or_unannounce(
-        &mut self,
-        keypair: &Keypair,
-        target: IdBytes,
-        token: &[u8; 32],
-        destination: PeerId,
-        relay_addresses: &[SocketAddr],
-        namespace: &[u8; 32],
-        // should be announce or unannounce
-        cmd: Command,
-    ) -> Tid {
-        let value = request_announce_or_unannounce_value(
-            keypair,
-            target,
-            token,
-            destination.id,
-            relay_addresses,
-            namespace,
-        );
-
-        let from_peer = Peer {
-            id: Some(destination.id.0),
-            addr: destination.addr,
-            referrer: None,
-        };
-
-        self.rpc.request(
-            OutRequestBuilder::new(from_peer, cmd)
-                .target(target)
-                .value(value)
-                .token(*token),
-        )
-    }
-
-    #[expect(unused)] // TODO FIXME
-    fn request_announce(
-        &mut self,
-        keypair: &Keypair,
-        target: IdBytes,
-        token: &[u8; 32],
-        destination: PeerId,
-        relay_addresses: &[SocketAddr],
-    ) -> Tid {
-        // TODO rm result
-        self.request_announce_or_unannounce(
-            keypair,
-            target,
-            token,
-            destination,
-            relay_addresses,
-            &crate::crypto::namespace::ANNOUNCE,
-            commands::ANNOUNCE,
-        )
-    }
-
-    #[expect(unused)] // TODO FIXME
-    fn request_unannounce(
-        &mut self,
-        keypair: &Keypair,
-        target: IdBytes,
-        token: &[u8; 32],
-        destination: PeerId,
-    ) -> Tid {
-        self.request_announce_or_unannounce(
-            keypair,
-            target,
-            token,
-            destination,
-            &[],
-            &crate::crypto::namespace::UNANNOUNCE,
-            commands::UNANNOUNCE,
-        )
-    }
-}
-
-impl Stream for HyperDhtInner {
-    type Item = HyperDhtEvent;
-
-    #[instrument(skip_all)]
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let pin = self.get_mut();
-
-        loop {
-            // Drain queued events first.
-            if let Some(event) = pin.queued_events.pop_front() {
-                cx.waker().wake_by_ref();
-                return Poll::Ready(Some(event));
-            }
-
-            // Drain rpc events
-            while let Poll::Ready(Some(ev)) = Stream::poll_next(Pin::new(&mut pin.rpc), cx) {
-                match ev {
-                    RpcDhtEvent::ResponseResult(Ok(ResponseOk::Response(resp))) => {
-                        _ = pin
-                            .inject_response(resp, cx)
-                            .inspect_err(|e| error!(error=?e,"failed to handle response"));
-                    }
-                    RpcDhtEvent::QueryResponse(resp) => {
-                        _ = pin
-                            .inject_query_response(resp, cx)
-                            .inspect_err(|e| error!(error=?e,"failed to handle response"));
-                    }
-                    RpcDhtEvent::Bootstrapped(bs) => {
-                        return Poll::Ready(Some(HyperDhtEvent::Bootstrapped(bs)));
-                    }
-                    RpcDhtEvent::RequestResult(Ok(RequestOk::CustomCommandRequest {
-                        query,
-                        request,
-                        peer,
-                    })) => pin.on_command(query, *request, peer),
-                    RpcDhtEvent::ReadyToCommit {
-                        query,
-                        tx_commit_messages,
-                    } => {
-                        pin.commit(query, tx_commit_messages);
-                    }
-                    RpcDhtEvent::QueryResult(qr) => {
-                        pin.query_target_search_done(qr);
-                    }
-                    RpcDhtEvent::RoutingUpdated { .. } => {
-                        // TODO
-                    }
-                    RpcDhtEvent::ResponseResult(Err(e)) => {
-                        error!(error=?e, "Got error response");
-                    }
-                    RpcDhtEvent::ResponseResult(Ok(ResponseOk::Pong(_))) => {}
-                    RpcDhtEvent::RequestResult(Err(e)) => {
-                        error!(error=?e, "Got error request");
-                    }
-                }
-            }
-
-            // Poll ongoing queries
-            while let Poll::Ready(Some(query_result)) = pin.queries.poll_next_unpin(cx) {
-                use HyperDhtEvent as hde;
-                use QueryStreamResult as qsr;
-                pin.queued_events.push_back(match query_result {
-                    Err(_) => todo!("we should just emit this as an error"),
-                    Ok(res) => match res {
-                        qsr::Lookup(r) => hde::LookupResult(r),
-                        qsr::Announce(r) => hde::AnnounceResult(r),
-                        qsr::UnAnnounce(r) => hde::UnAnnounceResult(Ok(r)),
-                        qsr::AnnounceClear(r) => hde::AnnouncClearResult(Ok(r)),
-                        qsr::FindPeer(r) => {
-                            let r = Arc::new(r);
-                            if let Some(tx) = pin.pending_find_peers.remove(&r.query_id) {
-                                _ = tx.send(r.clone());
-                            }
-                            hde::FindPeerResult(Ok(r))
-                        }
-                    },
-                })
-            }
-
-            // No immediate event was produced as a result of the DHT.
-            // If no new events have been queued either, signal `Pending` to
-            // be polled again later.
-            if pin.queued_events.is_empty() {
-                return Poll::Pending;
-            }
-        }
-    }
-}
-
-/// Future that resolves when a QueryResult is received for the find_peer query
-pub struct FindPeerFuture {
-    inner: Arc<Mutex<HyperDhtInner>>,
-    rx: oneshot::Receiver<Arc<QueryResult>>,
-}
-
-impl Future for FindPeerFuture {
-    type Output = Result<Arc<QueryResult>>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        {
-            let mut inner = self.inner.lock().unwrap();
-            let _ = Stream::poll_next(Pin::new(&mut *inner), cx);
-        }
-        Pin::new(&mut self.rx).poll(cx).map_err(Error::RecvError)
-    }
-}
-
-/// Events
-///
-/// The events produced by the `HyperDht` behaviour.
-///
-/// See [`HyperDht::poll`].
-// TODO should this be refactored into...
-// - enum {
-//   Response(ResponseKind) / Ann, Look, etc
-//   Result(ResultKind) / Ann, Look, etc
-//   ...
-// }
-// - enum {
-//   Announce(Ann/Result/Resp)
-//   Lookup(Ann/Result/Resp)
-// }
-// ... maybe for now just make it flat
-#[derive(Debug)]
-pub enum HyperDhtEvent {
-    /// The dht is now bootstrapped
-    Bootstrapped(Arc<Bootstrapped>),
-    /// The result of [`HyperDht::announce`].
-    AnnounceResult(QueryResult),
-    /// A response to part of a lookup query
-    LookupResponse(LookupResponse),
-    /// The result of [`HyperDht::lookup`].
-    LookupResult(QueryResult),
-    /// The result of [`HyperDht::unannounce`].
-    UnAnnounceResult(Result<UnannounceResult>),
-    /// The result of [`HyperDht::announce_clear`]
-    AnnouncClearResult(Result<AnnounceClearResult>),
-    /// A response to part of a find_peer query
-    FindPeerResponse(FindPeerResponse),
-    /// The result of [`HyperDht::find_peer`].
-    FindPeerResult(Result<Arc<QueryResult>>),
-    /// Received a query with a custom command that is not automatically handled
-    /// by the DHT
-    CustomCommandQuery {
-        /// The unknown command
-        command: usize,
-        /// The message we received from the peer.
-        msg: Box<RequestMsgData>,
-        /// The peer the message originated from.
-        peer: Peer,
-    },
-    Connected((Tid, Connection)),
-    PeerHandshakeResponse(Result<Arc<PeerHandshakeResponse>>),
-}
-
 #[derive(Debug)]
 pub struct PeerHandshakeResponse {
     noise: Vec<u8>,
@@ -784,24 +144,6 @@ impl PeerHandshakeResponse {
             relayed,
             server_address,
             client_address,
-        }
-    }
-}
-
-impl HyperDhtEvent {
-    pub fn kind(&self) -> &'static str {
-        match &self {
-            HyperDhtEvent::Bootstrapped { .. } => "Bootstrapped",
-            HyperDhtEvent::AnnounceResult(_) => "AnnounceResult",
-            HyperDhtEvent::LookupResponse(_) => "LookupResponse",
-            HyperDhtEvent::LookupResult(_) => "LookupResult",
-            HyperDhtEvent::UnAnnounceResult(_) => "UnAnnounceResult",
-            HyperDhtEvent::AnnouncClearResult(_) => "AnnouncClearResult",
-            HyperDhtEvent::FindPeerResponse(_) => "FindPeerResponse",
-            HyperDhtEvent::FindPeerResult(_) => "FindPeerResult",
-            HyperDhtEvent::CustomCommandQuery { .. } => "CustomCommandQuery",
-            HyperDhtEvent::Connected { .. } => "Connected",
-            HyperDhtEvent::PeerHandshakeResponse(_) => "PeerHandshakeResponse",
         }
     }
 }
@@ -837,139 +179,6 @@ pub struct Peers {
     pub peers: Vec<SocketAddr>,
     /// List of LAN peers that announced the topic hash
     pub local_peers: Vec<SocketAddr>,
-}
-
-/// Type to keep track of the responses for queries in progress.
-#[derive(Debug)]
-#[pin_project::pin_project(project = QueryStreamTypeProj)]
-enum QueryStreamType {
-    Lookup(LookupInner),
-    Announce(AnnounceInner),
-    UnAnnounce(#[pin] UnannounceInner),
-    AnnounceClear(AunnounceClearInner),
-    FindPeer(FindPeerInner),
-}
-
-#[derive(Debug)]
-enum QueryStreamResult {
-    Lookup(QueryResult),
-    Announce(QueryResult),
-    UnAnnounce(UnannounceResult),
-    AnnounceClear(AnnounceClearResult),
-    FindPeer(QueryResult),
-}
-
-impl Future for QueryStreamType {
-    type Output = Result<QueryStreamResult>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        use QueryStreamResult as qsr;
-        use QueryStreamTypeProj as qstp;
-        match self.project() {
-            qstp::UnAnnounce(mut inner) => {
-                if let Poll::Ready(x) = Future::poll(Pin::new(&mut inner), cx) {
-                    match x {
-                        Ok(res) => return Poll::Ready(Ok(qsr::UnAnnounce(res))),
-                        Err(e) => return Poll::Ready(Err(e)),
-                    }
-                }
-                Poll::Pending
-            }
-            qstp::Announce(mut inner) => {
-                if let Poll::Ready(x) = Future::poll(Pin::new(&mut inner), cx) {
-                    match x {
-                        Ok(res) => return Poll::Ready(Ok(qsr::Announce(res))),
-                        Err(e) => return Poll::Ready(Err(e)),
-                    }
-                }
-                Poll::Pending
-            }
-            qstp::Lookup(mut inner) => {
-                if let Poll::Ready(x) = Future::poll(Pin::new(&mut inner), cx) {
-                    match x {
-                        Ok(res) => return Poll::Ready(Ok(qsr::Lookup(res))),
-                        Err(e) => return Poll::Ready(Err(e)),
-                    }
-                }
-                Poll::Pending
-            }
-            qstp::AnnounceClear(mut inner) => {
-                if let Poll::Ready(x) = Future::poll(Pin::new(&mut inner), cx) {
-                    match x {
-                        Ok(res) => return Poll::Ready(Ok(qsr::AnnounceClear(res))),
-                        Err(e) => return Poll::Ready(Err(e)),
-                    }
-                }
-                Poll::Pending
-            }
-            qstp::FindPeer(mut inner) => {
-                if let Poll::Ready(x) = Future::poll(Pin::new(&mut inner), cx) {
-                    match x {
-                        Ok(res) => return Poll::Ready(Ok(qsr::FindPeer(res))),
-                        Err(e) => return Poll::Ready(Err(e)),
-                    }
-                }
-                Poll::Pending
-            }
-        }
-    }
-}
-
-impl QueryStreamType {
-    fn commit(&mut self, query: Arc<RwLock<Query>>, mut channel: mpsc::Sender<CommitMessage>) {
-        match self {
-            QueryStreamType::Lookup(_) => todo!(),
-            QueryStreamType::UnAnnounce(_) => todo!(),
-            QueryStreamType::Announce(inner) => {
-                let q = query.read().unwrap();
-                // TODO UGLY
-                warn!("# closest replies = [{}]", q.closest_replies.len());
-                for cr in q.closest_replies.iter() {
-                    let Some(pid) = cr.request.to.id else {
-                        // refactor this to be handled in the type system
-                        warn!("closest_replies peer without id.. Should not happen");
-                        continue;
-                    };
-                    trace!(
-                        "Sending commit to peer.id = [{:?}]",
-                        Into::<IdBytes>::into(pid)
-                    );
-                    channel
-                        .try_send(CommitMessage::Send(CommitRequestParams {
-                            command: commands::ANNOUNCE,
-                            target: Some(inner.topic),
-                            value: Some(request_announce_or_unannounce_value(
-                                &inner.keypair,
-                                inner.topic,
-                                &cr.response.token.expect("todo"),
-                                pid.into(),
-                                &[],
-                                &crate::crypto::namespace::ANNOUNCE,
-                            )),
-                            peer: cr.peer.addr,
-                            query_id: q.id,
-                            token: cr.response.token.expect("TODO"),
-                        }))
-                        .expect("TODO");
-                }
-
-                debug!("Emit CommitMessage::Done for query.id = {}", q.id);
-                channel.try_send(CommitMessage::Done).unwrap();
-            }
-            QueryStreamType::AnnounceClear(_) => todo!(),
-            QueryStreamType::FindPeer(_inner) => todo!(),
-        }
-    }
-
-    fn target_search_done(&mut self, msg_tx: MessageSender, query_result: Arc<RpcQueryResult>) {
-        match self {
-            QueryStreamType::Lookup(ref mut inner) => inner.finalize(),
-            QueryStreamType::Announce(ref mut inner) => inner.finalize(),
-            QueryStreamType::UnAnnounce(ref mut inner) => inner.finalize(),
-            QueryStreamType::AnnounceClear(ref mut inner) => inner.finalize(msg_tx, query_result),
-            QueryStreamType::FindPeer(ref mut inner) => inner.finalize(),
-        }
-    }
 }
 
 pub fn request_announce_or_unannounce_value(
